コメント記載とリファクタリング
